Value of $1 from 1800 to 2020
$1 in 1800 is equivalent in purchasing power to about $23.65 today, an increase of $22.65 over 223 years. The dollar had an average inflation rate of 1.43% per year between 1800 and today, producing a cumulative price increase of 2,265.16%.

This means that today's prices are 23.65 times as high as average prices since 1800, according to the Bureau of Labor Statistics consumer price index. A dollar today only buys 4.228% of what it could buy back then.

The inflation rate in 1800 was 2.44%. The current inflation rate compared to last year is now 7.11%. If this number holds, $1 today will be equivalent in buying power to $1.07 next year. The current inflation rate page gives more detail on the latest inflation rates.